Added two visualization prototypes.
Used JUNG and JGraph libraries to see which one would be better for the
purposes of solving my “harder” problem for this project.  This is for
the “easy” problem of the project which is to add a way to visualize
the social network graph.  Next step is to implement a network cascade
algorithm.
-Also commented out some code I wasn’t using, such as the grader
classes.